requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
opencv-python==4.8.1.78
psycopg2-binary==2.9.9
transformers==4.35.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.1+cpu
tensorflow==2.16.1
scikit-learn==1.3.2
xgboost==2.0.2
spacy==3.6.1
openai==1.3.5
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
en-core-sci-scibert @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.3/en_core_sci_scibert-0.5.3.tar.gz 